requirements.txt
dash==2.14.2
plotly==5.18.0
numpy==1.24.3
Pillow==10.1.0

